requirements.txt
setuptools==58.0.4
opencv-python==4.5.5.64
opencv-contrib-python
autopep8
einops==0.6.0
h5py==3.8.0
loguru==0.5.3
matplotlib==3.5.1
pylint==2.17.1
ray>=1.0.1
torchmetrics==0.6.0
tqdm==4.64.1
yacs>=0.1.8
joblib>=1.0.1
numpy==1.23.2